# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Almera E-Commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/orders:
    get:
      operationId: adminListOrders
      tags: [admin]
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Paginated orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedOrders"
  /admin/orders/{id}:
    patch:
      operationId: adminUpdateOrder
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderRequest"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /admin/settings:
    put:
      operationId: adminUpdateSettings
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: string
      responses:
        "200":
          description: Settings updated
...
  /admin/legal/{id}:
    put:
      operationId: adminUpdateLegalPage
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLegalPageRequest"
      responses:
        "200":
          description: Updated legal page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegalPage"
  /admin/products:
    get:
      operationId: adminListProducts
      tags: [admin]
      responses:
        "200":
          description: All products for admin
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: adminCreateProduct
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpsertRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /admin/products/{id}:
    put:
      operationId: adminUpdateProduct
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpsertRequest"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: adminDeleteProduct
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
  /admin/upload/product-image:
    post:
      operationId: adminUploadProductImage
      tags: [admin]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Uploaded image URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
...
  /admin/reviews/{id}:
    delete:
      operationId: adminDeleteReview
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registered and logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionUser"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionUser"
...
components:
  schemas:
...
    UploadResponse:
      type: object
      properties:
        url: { type: string }
      required: [url]
    ProductVariant:
      type: object
      properties:
        size: { type: string }
        color: { type: string }
        quality: { type: string }
        stock: { type: integer }
      required: [size, color, quality, stock]
    Product:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        description: { type: string }
        category: { type: string }
        originalPrice: { type: number }
        discountedPrice: { type: number }
        images:
          type: array
          items: { type: string }
        videos:
          type: array
          items: { type: string }
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariant"
        rating: { type: number }
        reviewCount: { type: integer }
        isFeatured: { type: boolean }
        isNew: { type: boolean }
      required: [id, name, slug, description, category, originalPrice, discountedPrice, images, variants, rating, reviewCount, isFeatured, isNew]
    ProductUpsertRequest:
      type: object
      properties:
        name: { type: string }
        slug: { type: string }
        description: { type: string }
        category: { type: string }
        originalPrice: { type: number }
        discountedPrice: { type: number }
        images:
          type: array
          items: { type: string }
        videos:
          type: array
          items: { type: string }
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariant"
        isFeatured: { type: boolean }
        isNew: { type: boolean }
      required: [name, slug, description, category, originalPrice, discountedPrice]
...
    OrderItem:
      type: object
      properties:
        productId: { type: string }
        productName: { type: string }
        productImage:
          type: string
          nullable: true
        size: { type: string }
        color: { type: string }
        quantity: { type: integer }
        unitPrice: { type: number }
        totalPrice: { type: number }
      required: [productId, productName, size, color, quantity, unitPrice, totalPrice]
...
    Order:
      type: object
      properties:
        id: { type: string }
        orderNumber: { type: string }
        customerName: { type: string }
        customerEmail:
          type: string
          nullable: true
        customerPhone: { type: string }
        shippingAddress: { type: string }
        shippingCity: { type: string }
        paymentMethod: { type: string }
        paymentScreenshotUrl:
          type: string
          nullable: true
        paymentStatus: { type: string }
        status: { type: string }
        subtotal: { type: number }
        deliveryFee: { type: number }
        codFee: { type: number }
        total: { type: number }
        notes:
          type: string
          nullable: true
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, orderNumber, customerName, customerPhone, shippingAddress, shippingCity, paymentMethod, paymentStatus, status, subtotal, deliveryFee, codFee, total, createdAt, updatedAt]
    OrderWithItems:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/OrderItem"
          required: [items]
    UpdateOrderRequest:
      type: object
      properties:
        status: { type: string }
        paymentStatus: { type: string }
    PaginatedOrders:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/OrderWithItems"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
      required: [data, total, page, limit]
...
    LegalPage:
      type: object
      properties:
        id: { type: string }
        slug: { type: string }
        title: { type: string }
        content: { type: string }
        updatedAt: { type: string }
      required: [id, slug, title, content, updatedAt]
    UpdateLegalPageRequest:
      type: object
      properties:
        title: { type: string }
        content: { type: string }
    SessionUser:
      type: object
      properties:
        id: { type: string }
        email:
          type: string
          nullable: true
        fullName:
          type: string
          nullable: true
        role: { type: string }
        isLoggedIn: { type: boolean }
      required: [id, role, isLoggedIn]
    RegisterRequest:
      type: object
      properties:
        email: { type: string }
        password: { type: string }
        fullName: { type: string }
      required: [email, password]
    LoginRequest:
      type: object
      properties:
        email: { type: string }
        password: { type: string }
      required: [email, password]