# lib/api-zod/src/generated/types/legalPage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Almera E-Commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface LegalPage {
  id: string;
  slug: string;
  title: string;
  content: string;
  updatedAt: string;
}