# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Almera E-Commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface Order {
  id: string;
  orderNumber: string;
  customerName: string;
  customerEmail?: string | null;
  customerPhone: string;
  shippingAddress: string;
  shippingCity: string;
  paymentMethod: string;
  paymentScreenshotUrl?: string | null;
  paymentStatus: string;
  status: string;
  subtotal: number;
  deliveryFee: number;
  codFee: number;
  total: number;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}