# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Almera E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductVariant } from "./productVariant";

export interface Product {
  id: string;
  name: string;
  slug: string;
  description: string;
  category: string;
  originalPrice: number;
  discountedPrice: number;
  images: string[];
  videos?: string[];
  variants: ProductVariant[];
  rating: number;
  reviewCount: number;
  isFeatured: boolean;
  isNew: boolean;
}