# lib/api-zod/src/generated/types/sessionUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Almera E-Commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface SessionUser {
  id: string;
  email?: string | null;
  fullName?: string | null;
  role: string;
  isLoggedIn: boolean;
}